permissions for routes done

+++ controllers/coursesController.js
@@ -40,10 +40,17 @@ export const getCourse = asyncHandler(async (req, res, next) => {
 // @access  Private
 export const createCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId
+    req.body.user = req.user.id
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.bootcampId}`, 404))
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${req.params.bootcampId}`, 401))
+    }
+
     const course = await Course.create(req.body)
     res.status(201).json({ success: true, data: course })
 })
@@ -53,13 +60,21 @@ export const createCourse = asyncHandler(async (req, res, next) => {
 // @access  Private
 export const updateCourse = asyncHandler(async (req, res, next) => {
     const { id } = req.params
-    const updatedCourse = await Course.findByIdAndUpdate(id, req.body, {
-        new: true,
-        runValidators: true,
-    })
+    let updatedCourse = await Course.findById(id)
     if (!updatedCourse) {
         return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404))
     }
+
+    // Make sure user is bootcamp owner
+    if (updatedCourse.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course`, 401))
+    }
+
+    updateCourse = await Course.findByIdAndUpdate(id, req.body, {
+        new: true,
+        runValidators: true,
+    })
+
     res.status(200).json({ success: true, data: updatedCourse })
 })
 
@@ -72,6 +87,12 @@ export const deleteCourse = asyncHandler(async (req, res, next) => {
     if (!removedCourse) {
         return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404))
     }
+
+    // Make sure user is bootcamp owner
+    if (removedCourse.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course`, 401))
+    }
+
     await removedCourse.deleteOne()
     res.status(200).json({ success: true, data: removedCourse })
 })
